# gateway/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@erp-system.com
    name: ERP System Support
  description: Enterprise Resource Planning System API Gateway. Provides unified REST
    API access to Inventory, Order, Finance, and HR microservices.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ERP System API
  version: "1.0"
paths:
...
  /api/v1/orders/customers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all customers
      tags:
      - Orders - Customers
    post:
      consumes:
      - application/json
      description: Create a new customer record
      parameters:
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new customer
      tags:
      - Orders - Customers
  /api/v1/orders/customers/{id}:
    get:
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a customer by ID
      tags:
      - Orders - Customers
  /api/v1/orders/purchase:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all purchase orders
      tags:
      - Orders - Purchase Orders
    post:
      consumes:
      - application/json
      parameters:
      - description: PO data
        in: body
        name: order
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a purchase order
      tags:
      - Orders - Purchase Orders
  /api/v1/orders/purchase/{id}:
    get:
      parameters:
      - description: PO ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a purchase order by ID
      tags:
      - Orders - Purchase Orders
  /api/v1/orders/purchase/{id}/receive:
    post:
      consumes:
      - application/json
      parameters:
      - description: PO ID
        in: path
        name: id
        required: true
        type: string
      - description: Receipt data
        in: body
        name: receipt
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Receive a purchase order
      tags:
      - Orders - Purchase Orders
  /api/v1/orders/sales:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all sales orders
      tags:
      - Orders - Sales Orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new sales order
      tags:
      - Orders - Sales Orders
  /api/v1/orders/sales/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a sales order by ID
      tags:
      - Orders - Sales Orders
  /api/v1/orders/sales/{id}/approve:
    post:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Approve a sales order
      tags:
      - Orders - Sales Orders
  /api/v1/orders/sales/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - Orders - Sales Orders
  /api/v1/orders/summary:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order summary dashboard
      tags:
      - Orders - Sales Orders
  /api/v1/orders/suppliers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all suppliers
      tags:
      - Orders - Suppliers
    post:
      consumes:
      - application/json
      parameters:
      - description: Supplier data
        in: body
        name: supplier
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new supplier
      tags:
      - Orders - Suppliers
  /api/v1/orders/suppliers/{id}:
    get:
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a supplier by ID
      tags:
      - Orders - Suppliers
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGciOiJI..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...