# gateway/handlers/order.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"google.golang.org/grpc"
)

type OrderHandler struct {
	conn *grpc.ClientConn
}

func NewOrderHandler(conn *grpc.ClientConn) *OrderHandler {
	return &OrderHandler{conn: conn}
}

// CreateCustomer godoc
// @Summary      Create a new customer
// @Description  Create a new customer record
// @Tags         Orders - Customers
// @Accept       json
// @Produce      json
// @Param        customer  body      map[string]interface{}  true  "Customer data"
// @Success      201       {object}  map[string]interface{}
// @Failure      400       {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/customers [post]
func (h *OrderHandler) CreateCustomer(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Customer created", "data": req})
}

// GetCustomer godoc
// @Summary      Get a customer by ID
// @Tags         Orders - Customers
// @Produce      json
// @Param        id   path      string  true  "Customer ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/customers/{id} [get]
func (h *OrderHandler) GetCustomer(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListCustomers godoc
// @Summary      List all customers
// @Tags         Orders - Customers
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/customers [get]
func (h *OrderHandler) ListCustomers(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List customers"}) }

// CreateOrder godoc
// @Summary      Create a new sales order
// @Tags         Orders - Sales Orders
// @Accept       json
// @Produce      json
// @Param        order  body      map[string]interface{}  true  "Order data"
// @Success      201    {object}  map[string]interface{}
// @Failure      400    {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/sales [post]
func (h *OrderHandler) CreateOrder(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	req["created_by"], _ = c.Get("userID")
	c.JSON(http.StatusCreated, gin.H{"message": "Order created", "data": req})
}

// GetOrder godoc
// @Summary      Get a sales order by ID
// @Tags         Orders - Sales Orders
// @Produce      json
// @Param        id   path      string  true  "Order ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/sales/{id} [get]
func (h *OrderHandler) GetOrder(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListOrders godoc
// @Summary      List all sales orders
// @Tags         Orders - Sales Orders
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/sales [get]
func (h *OrderHandler) ListOrders(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List orders"}) }

// UpdateOrderStatus godoc
// @Summary      Update order status
// @Tags         Orders - Sales Orders
// @Accept       json
// @Produce      json
// @Param        id      path      string                 true  "Order ID"
// @Param        status  body      map[string]interface{}  true  "New status"
// @Success      200     {object}  map[string]interface{}
// @Failure      400     {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/sales/{id}/status [put]
func (h *OrderHandler) UpdateOrderStatus(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"id": c.Param("id"), "data": req})
}

// ApproveOrder godoc
// @Summary      Approve a sales order
// @Tags         Orders - Sales Orders
// @Produce      json
// @Param        id   path      string  true  "Order ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/sales/{id}/approve [post]
func (h *OrderHandler) ApproveOrder(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"message": "Order approved", "id": c.Param("id")})
}

// CreatePurchaseOrder godoc
// @Summary      Create a purchase order
// @Tags         Orders - Purchase Orders
// @Accept       json
// @Produce      json
// @Param        order  body      map[string]interface{}  true  "PO data"
// @Success      201    {object}  map[string]interface{}
// @Failure      400    {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/purchase [post]
func (h *OrderHandler) CreatePurchaseOrder(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "PO created", "data": req})
}

// GetPurchaseOrder godoc
// @Summary      Get a purchase order by ID
// @Tags         Orders - Purchase Orders
// @Produce      json
// @Param        id   path      string  true  "PO ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/purchase/{id} [get]
func (h *OrderHandler) GetPurchaseOrder(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListPurchaseOrders godoc
// @Summary      List all purchase orders
// @Tags         Orders - Purchase Orders
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/purchase [get]
func (h *OrderHandler) ListPurchaseOrders(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List POs"}) }

// ReceivePurchaseOrder godoc
// @Summary      Receive a purchase order
// @Tags         Orders - Purchase Orders
// @Accept       json
// @Produce      json
// @Param        id       path      string                 true  "PO ID"
// @Param        receipt  body      map[string]interface{}  true  "Receipt data"
// @Success      200      {object}  map[string]interface{}
// @Failure      400      {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/purchase/{id}/receive [post]
func (h *OrderHandler) ReceivePurchaseOrder(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "PO received", "id": c.Param("id"), "data": req})
}

// CreateSupplier godoc
// @Summary      Create a new supplier
// @Tags         Orders - Suppliers
// @Accept       json
// @Produce      json
// @Param        supplier  body      map[string]interface{}  true  "Supplier data"
// @Success      201       {object}  map[string]interface{}
// @Failure      400       {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/suppliers [post]
func (h *OrderHandler) CreateSupplier(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Supplier created", "data": req})
}

// GetSupplier godoc
// @Summary      Get a supplier by ID
// @Tags         Orders - Suppliers
// @Produce      json
// @Param        id   path      string  true  "Supplier ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/suppliers/{id} [get]
func (h *OrderHandler) GetSupplier(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListSuppliers godoc
// @Summary      List all suppliers
// @Tags         Orders - Suppliers
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/suppliers [get]
func (h *OrderHandler) ListSuppliers(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List suppliers"}) }

// GetOrderSummary godoc
// @Summary      Get order summary dashboard
// @Tags         Orders - Sales Orders
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/orders/summary [get]
func (h *OrderHandler) GetOrderSummary(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"message": "Order summary"})
}